// client/swagger/model/Topic.ts
/**
 * LearnIt
 * Temporal Api for LearnIt.
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Topic {
    id: string;

    bundleId: string;

    languageId: string;

    name: string;

    cards: Array<models.Card>;

    waitingCards: Array<models.Card>;

    customTopic?: boolean;

    csvStringUrl?: string;

}